Fix your ssh key loading
The `if` command takes the status code of the command that it executes as the true or false value. So you don't need to test the status code of the previous command - just make that the command if runs.

The find command supports full boolean set logic including negation so there is no need to chain a grep off that. Also chaining for and the associated command on one line doesn't read well.